Fix issue 321, prevent placeCOP from being issuable an infinite number of times. Fix a bug that restricted placement of an Infested Command Center and Zerg Flag Beacon outside of the creep.
git-svn-id: https://bwapi.googlecode.com/svn/trunk@3258 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAPI/Source/BW/OrderTypes.cpp
@@ -196,7 +196,7 @@ namespace BW
         : always0x0c(0x0C)
         , always0x9B(0x9B)
         , position(position)
-        , type(type.getID())
+        , type((u16)type.getID())
     {
     }
     //--------------------------------------------- INVENT TECH ----------------------------------------------
